+++ packages/test/src/m3/types.test.ts
@@ -1,39 +1,31 @@
-// const { throws } = require("chai").assert
-// import {LanguageFactory} from "../../src-pkg/m3/factory.js"
-// import {builtinPrimitives} from "../../src-pkg/m3/builtins.js"
+import {assert} from "chai"
+const {equal} = assert
 
-
-/*
- * The following typechecks because all classes are effectively one type "class" as TypeScript is concerned.
- */
-class A {
-    with(_: _B) {}
-}
-class _B {}
-const a = new A()
-a.with(a)   // would be nice if this produced a compiler error!
-/*
- * This means that we have to check arguments of a class-type "manually".
- */
+import {concatenator, LanguageFactory, lastOf} from "@lionweb/core"
 
 
 describe("M3 types", () => {
 
-    /*
-    it("adding non-language elements to a language should fail", () => {
-        const factory = new LanguageFactory("TestLanguage", "0")
+    it("factory performs auto-containment, but also prevents duplication", () => {
+        const factory = new LanguageFactory("TestLanguage", "0", concatenator("-"), lastOf)
         const {language} = factory
+
         const concept = factory.concept("Concept", false)
-        const property = factory.property(concept, "property").ofType(builtinPrimitives.intDatatype)
-        factory.language.havingEntities(concept)
-        // TODO  understand why the following fails to compile after the addition of a getter "language" to LanguageEntity
-        throws(() => {
-            language.havingEntities(property),
-            Error,
-            `trying to add non-LanguageElements to Language: <Property>"property"`
-        })
+        equal(language.entities.length, 1)
+        language.havingEntities(concept)
+        equal(language.entities.length, 1)
+
+        const containment = factory.containment(concept, "containment")
+        equal(concept.features.length, 1)
+        concept.havingFeatures(containment)
+        equal(concept.features.length, 1)
+
+        const enumeration = factory.enumeration("enumeration")
+        const literal = factory.enumerationLiteral(enumeration, "literal")
+        equal(enumeration.literals.length, 1)
+        enumeration.havingLiterals(literal)
+        equal(enumeration.literals.length, 1)
     })
-     */
 
 })